Backdrop Generator Wizard adjustments
